Refactor C2H headers by responsibility
